requirements.txt
Flask>=2.1.1
openai==0.27.0
python-dotenv==0.19.2
werkzeug>=2.0.0